requirements.txt
lxml
requests
python-dateutil
fastapi
uvicorn
pyyaml
